move MethodCallToAnotherMethodCallWithArgumentsRector form Nette to generic Rector/MethodCall
