--- packages/Nette/tests/Rector/DI/MethodCallToAnotherMethodCallWithArgumentsRector/Source/NetteServiceDefinition.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Nette\Tests\Rector\DI\MethodCallToAnotherMethodCallWithArgumentsRector\Source;
-
-final class NetteServiceDefinition
-{
-
-}
