+++ tests/Rector/MethodCall/MethodCallToAnotherMethodCallWithArgumentsRector/Correct/correct.php.inc
