+++ tests/Rector/MethodCall/MethodCallToAnotherMethodCallWithArgumentsRector/Source/NetteServiceDefinition.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\MethodCall\MethodCallToAnotherMethodCallWithArgumentsRector\Source;
+
+final class NetteServiceDefinition
+{
+
+}
