+++ tests/Rector/MethodCall/MethodCallToAnotherMethodCallWithArgumentsRector/Wrong/wrong.php.inc
